refactor page name
