+++ tourou/test/ui/pages/goldens/new_profile_setting_page_iphone6_ja.png
