+++ tourou/test/ui/pages/new_profile_setting_page_widget_test.dart
@@ -8,7 +8,7 @@ import 'package:golden_toolkit/golden_toolkit.dart';
 
 // Project imports:
 import '../../../lib/l10n/app_localizations.dart';
-import '../../../lib/ui/pages/new_user_registration_page.dart';
+import '../../../lib/ui/pages/new_profile_setting_page.dart';
 
 Future<void> loadJapaneseFont() async {
   TestWidgetsFlutterBinding.ensureInitialized();
@@ -27,7 +27,7 @@ Widget TestWidget(String language) {
 }
 
 void main() {
-  testGoldens('new_user_registration_page_golden_test',
+  testGoldens('new_profile_setting_page_golden_test',
       (WidgetTester tester) async {
     await loadAppFonts();
     await loadJapaneseFont();
@@ -39,13 +39,13 @@ void main() {
     await tester.pumpWidgetBuilder(TestWidget('en'), surfaceSize: size6);
 
     //マスターのスクリーンショットと同じかテストする
-    await screenMatchesGolden(tester, 'new_user_registration_page_iphone6_en');
+    await screenMatchesGolden(tester, 'new_profile_setting_page_iphone6_en');
 
     //第一引数はどのWidgetをビルドするのか指定、どのサイズにビルドするかがsurfaceSize
     await tester.pumpWidgetBuilder(TestWidget('ja'), surfaceSize: size6);
 
     //マスターのスクリーンショットと同じかテストする
-    await screenMatchesGolden(tester, 'new_user_registration_page_iphone6_ja');
+    await screenMatchesGolden(tester, 'new_profile_setting_page_iphone6_ja');
 
     //デバイスの画面サイズ
     final sizePad = Size(1024, 1366);
@@ -54,12 +54,12 @@ void main() {
     await tester.pumpWidgetBuilder(TestWidget('en'), surfaceSize: sizePad);
 
     //マスターのスクリーンショットと同じかテストする
-    await screenMatchesGolden(tester, 'new_user_registration_page_ipad_en');
+    await screenMatchesGolden(tester, 'new_profile_setting_page_ipad_en');
 
     //第一引数はどのWidgetをビルドするのか指定、どのサイズにビルドするかがsurfaceSize
     await tester.pumpWidgetBuilder(TestWidget('ja'), surfaceSize: sizePad);
 
     //マスターのスクリーンショットと同じかテストする
-    await screenMatchesGolden(tester, 'new_user_registration_page_ipad_ja');
+    await screenMatchesGolden(tester, 'new_profile_setting_page_ipad_ja');
   });
 }
